Move PublicHolidaysDataTable class to modules and inport it into the main-v1.js file
