+++ unit-13/exercise-3-class-public-holidays-table/index.html
@@ -17,7 +17,7 @@
         <small>J.N &copy 2025</small>
       </section>
     </footer>
-    <!-- <script src="js/main-v1.js"></script> -->
-    <script src="js/main-v2.js"></script>
+    <script type="module" src="js/main-v1.js"></script>
+    <!-- <script src="js/main-v2.js"></script> -->
   </body>
 </html>
